Hit an API, added a carousel and a skeleton state.
